refactor: modularize exercise library page
