+++ src/pages/exercise-library/constants.ts
@@ -1,0 +1,153 @@
+import type {TranslationRequest} from '@/types/translations'
+
+export const EXERCISE_LIBRARY_TRANSLATIONS: TranslationRequest[] = [
+	{
+		key: 'exerciseLibrary',
+		fallback: 'Exercise Library'
+	},
+	{
+		key: 'exerciseLibraryDesc',
+		fallback: 'Browse and execute available exercises'
+	},
+	{
+		key: 'settings',
+		fallback: 'Settings'
+	},
+	{
+		key: 'hintLanguage',
+		fallback: 'Hint language'
+	},
+	{
+		key: 'userLanguageDescription',
+		fallback:
+			'Choose a language you already know. It will be used for hints in exercises.'
+	},
+	{
+		key: 'exerciseCount',
+		fallback: 'Available exercises: {filteredCount}'
+	},
+	{
+		key: 'noExercisesFound',
+		fallback: 'No exercises found'
+	},
+	{
+		key: 'noExercisesFoundDesc',
+		fallback: 'Try adjusting your filters or browse all exercises'
+	},
+	{
+		key: 'clearFilters',
+		fallback: 'Clear Filters'
+	},
+	{
+		key: 'startExercise',
+		fallback: 'Start Exercise'
+	},
+	{
+		key: 'learn',
+		fallback: 'Learn'
+	},
+	{
+		key: 'filters',
+		fallback: 'Filters'
+	},
+	{
+		key: 'difficulty',
+		fallback: 'Difficulty'
+	},
+	{
+		key: 'tags',
+		fallback: 'Tags'
+	},
+	{
+		key: 'all',
+		fallback: 'All'
+	},
+	{
+		key: 'ui.searchEmoji',
+		fallback: '🔍'
+	},
+	{
+		key: 'ui.documentEmoji',
+		fallback: '📝'
+	},
+	{
+		key: 'ui.booksEmoji',
+		fallback: '📚'
+	},
+	{
+		key: 'ui.timerEmoji',
+		fallback: '⏱️'
+	},
+	{
+		key: 'exercise.cases',
+		fallback: 'cases'
+	},
+	{
+		key: 'exercise.blocks',
+		fallback: 'blocks'
+	},
+	{
+		key: 'exercise.minutes',
+		fallback: 'min'
+	},
+	{
+		key: 'ui.hashSymbol',
+		fallback: '#'
+	},
+	{
+		key: 'ui.plusSymbol',
+		fallback: '+'
+	},
+	{
+		key: 'ui.hintLanguage',
+		fallback: 'Hint language'
+	},
+	{
+		key: 'ui.chevronDown',
+		fallback: '▼'
+	},
+	{
+		key: 'ui.colon',
+		fallback: ':'
+	},
+	{
+		key: 'ui.chevronUp',
+		fallback: '▲'
+	},
+	{
+		key: 'ui.expand',
+		fallback: 'Expand'
+	},
+	{
+		key: 'ui.collapse',
+		fallback: 'Collapse'
+	},
+	{
+		key: 'difficulty.a0',
+		fallback: 'A0'
+	},
+	{
+		key: 'difficulty.a1',
+		fallback: 'A1'
+	},
+	{
+		key: 'difficulty.a2',
+		fallback: 'A2'
+	},
+	{
+		key: 'difficulty.b1',
+		fallback: 'B1'
+	},
+	{
+		key: 'difficulty.b2',
+		fallback: 'B2'
+	},
+	{
+		key: 'difficulty.c1',
+		fallback: 'C1'
+	},
+	{
+		key: 'difficulty.c2',
+		fallback: 'C2'
+	}
+]
